requirements.txt
streamlit
pandas
numpy
nibabel
matplotlib
torch
torchvision
Pillow
albumentations
google-generativeai